refactor: Auth 도메인 관련 DTO 파일 리패키징
