+++ backend/src/main/java/com/icehufs/icebreaker/domain/auth/dto/request/EmailCertificationRequestDto.java
@@ -1,4 +1,4 @@
-package com.icehufs.icebreaker.dto.request.auth;
+package com.icehufs.icebreaker.domain.auth.dto.request;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
